Small fixes
1) Moved _matplotlib_list to matplotlib backend
2) Renamed get_meshes() in plot_implicit to get_raster
3) Improved the documentation
4) Dummy used when there are less than 2 free variables.
